Fix Windows compiler warnings
Now that we have regular Windows builds in the GitHub actions, with
warnings as errors, these can get fixed and stay fixed.

Diffs=
25bc5535c Fix Windows compiler warnings

+++ src/animation/linear_animation_instance.cpp
@@ -65,13 +65,13 @@ bool LinearAnimationInstance::advance(float elapsedSeconds) {
             if (m_Direction == 1 && frames > end) {
                 keepGoing = false;
                 m_SpilledTime = (frames - end) / fps;
-                frames = end;
+                frames = (float)end;
                 m_Time = frames / fps;
                 didLoop = true;
             } else if (m_Direction == -1 && frames < start) {
                 keepGoing = false;
                 m_SpilledTime = (start - frames) / fps;
-                frames = start;
+                frames = (float)start;
                 m_Time = frames / fps;
                 didLoop = true;
             }
